+++ src/views/base/users/Groups.js
@@ -39,6 +39,7 @@ const Groups = () => {
     const [newGroupModalVisible, setNewGroupModalVisible] = useState(false);
     const [newGroupName, setNewGroupName] = useState('');
     const [newGroupUsers, setNewGroupUsers] = useState([]);
+    const [successMessage, setSuccessMessage] = useState('');
 
     const getAuthHeaders = () => {
         const token = localStorage.getItem('token');
@@ -73,14 +74,13 @@ const Groups = () => {
         fetchUserData();
     }, [fetchGroupData, fetchUserData]);
 
-    const handleSearch = useCallback((query) => {
-        setSearchQuery(query);
+    const handleSearch = useCallback(() => {
         if (!groupData) return;
 
-        if (query === '') {
+        if (searchQuery === '') {
             setFilteredGroupData(groupData);
         } else {
-            const lowercasedQuery = query.toLowerCase();
+            const lowercasedQuery = searchQuery.toLowerCase();
             const filteredData = groupData.filter(group => {
                 const name = group.name?.toLowerCase() || '';
                 const userSet = group.user_set ? group.user_set.map(userId => userId.toString().toLowerCase()).join(' ') : '';
@@ -96,11 +96,7 @@ const Groups = () => {
             });
             setFilteredGroupData(filteredData);
         }
-    }, [groupData]);
-
-    useEffect(() => {
-        handleSearch(searchQuery);
-    }, [handleSearch, searchQuery]);
+    }, [groupData, searchQuery]);
 
     const handleGroupSelect = (group) => {
         if (group) {
@@ -136,17 +132,25 @@ const Groups = () => {
         setNewGroupUsers(selectedUsers);
     };
 
+    const showSuccessMessage = (message) => {
+        setSuccessMessage(message);
+        setTimeout(() => setSuccessMessage(''), 3000);
+    };
+
     const handleUpdateGroup = async () => {
         try {
             if (selectedGroup.id) {
                 await axios.put(`${BaseURL}app/groups/${selectedGroup.id}/`, selectedGroup, { headers: getAuthHeaders() });
+                showSuccessMessage('Group updated successfully!');
             } else {
                 await axios.post(`${BaseURL}app/groups/`, selectedGroup, { headers: getAuthHeaders() });
+                showSuccessMessage('Group created successfully!');
             }
             fetchGroupData();
             setModalVisible(false);
         } catch (error) {
             console.error('Error updating group:', error);
+            showSuccessMessage('Error updating group!');
         }
     };
 
@@ -157,31 +161,52 @@ const Groups = () => {
                 user_set: newGroupUsers,
             };
             await axios.post(BaseURL + "app/groups/", newGroup, { headers: getAuthHeaders() });
+            showSuccessMessage('New group created successfully!');
             fetchGroupData();
             setNewGroupModalVisible(false);
             setNewGroupName('');
             setNewGroupUsers([]);
         } catch (error) {
             console.error('Error creating new group:', error);
+            showSuccessMessage('Error creating new group!');
         }
     };
 
     const handleDeleteGroup = async (groupId) => {
+        const confirmDelete = window.confirm("Are you sure you want to delete this group?");
+        
+        if (!confirmDelete) return;
+
         try {
             await axios.delete(`${BaseURL}app/groups/${groupId}/`, { headers: getAuthHeaders() });
+            showSuccessMessage('Group deleted successfully!');
             fetchGroupData();
         } catch (error) {
             console.error('Error deleting group:', error);
+            showSuccessMessage('Error deleting group!');
         }
     };
 
     return (
         <>
+            {successMessage && (
+                <div className="alert alert-success" role="alert">
+                    {successMessage}
+                </div>
+            )}
+
             <CRow>
                 <CCol xs={12}>
                     <CCard className="mb-4">
                         <CCardHeader>
-                            <strong>Group LIST</strong>
+                            <div className="d-flex align-items-center justify-content-between">
+                                <strong>Group LIST</strong>
+                                <CTooltip content="Create new Group">
+                                    <CButton type="button" color="success" variant='outline' size='sm' onClick={() => setNewGroupModalVisible(true)}>
+                                        Create Group
+                                    </CButton>
+                                </CTooltip>
+                            </div>
                         </CCardHeader>
                         <CCardBody>
                             <CCol md={4}>
@@ -191,18 +216,13 @@ const Groups = () => {
                                         aria-label="Search"
                                         aria-describedby="addon-wrapping"
                                         value={searchQuery}
-                                        onChange={(e) => handleSearch(e.target.value)}
+                                        onChange={(e) => setSearchQuery(e.target.value)}
                                     />
-                                    <CButton type="button" color="secondary" onClick={(e) => handleSearch(e.target.value)} id="button-addon2">
+                                    <CButton type="button" color="secondary" onClick={handleSearch} id="button-addon2">
                                         Search
                                     </CButton>
                                 </CInputGroup>
                             </CCol>
-                            <CTooltip content="Create new Group">
-                                <CButton type="button" color="primary" className="mb-3" onClick={() => setNewGroupModalVisible(true)}>
-                                    Create
-                                </CButton>
-                            </CTooltip>
                             <CTable striped hover>
                                 <CTableHead>
                                     <CTableRow color="dark">
@@ -226,14 +246,12 @@ const Groups = () => {
                                                 <CTableDataCell>{group.user_details.map(user => user.email).join(', ')}</CTableDataCell>
                                                 <CTableDataCell>{group.user_details.map(user => user.mobile_no).join(', ')}</CTableDataCell>
                                                 <CTableDataCell>
-                                                    <div className="d-flex gap-2">
-                                                        <CButton onClick={() => handleGroupSelect(group)}>
-                                                            <CIcon icon={cilPen} />
-                                                        </CButton>
-                                                        <CButton onClick={(e) => { e.stopPropagation(); handleDeleteGroup(group.id); }}>
-                                                            <CIcon icon={cilTrash} />
-                                                        </CButton>
-                                                    </div>
+                                                    <CButton color="primary" size='sm' onClick={() => handleGroupSelect(group)}>
+                                                        <CIcon icon={cilPen} />
+                                                    </CButton>
+                                                    <CButton color="primary" size='sm' onClick={() => handleDeleteGroup(group.id)} className="ms-2">
+                                                        <CIcon icon={cilTrash} />
+                                                    </CButton>
                                                 </CTableDataCell>
                                             </CTableRow>
                                         ))
@@ -244,78 +262,70 @@ const Groups = () => {
                     </CCard>
                 </CCol>
             </CRow>
-            <CModal size='lg' visible={modalVisible} onClose={() => setModalVisible(false)} backdrop="static" keyboard={false}>
-                <CModalHeader onClose={() => setModalVisible(false)}>
-                    <CModalTitle>Groups Details</CModalTitle>
+            <CModal visible={modalVisible} onClose={() => setModalVisible(false)}>
+                <CModalHeader>
+                    <CModalTitle>{selectedGroup && selectedGroup.id ? 'Edit Group' : 'New Group'}</CModalTitle>
                 </CModalHeader>
                 <CModalBody>
                     <CForm>
-                        <CRow className="mb-3">
-                            <CFormLabel htmlFor="group" className="col-sm-2 col-form-label">Group</CFormLabel>
-                            <CCol md={6}>
-                                <CFormInput type="text" id="group" name="group" value={selectedGroup ? selectedGroup.name : ''} readOnly />
-                            </CCol>
-                        </CRow>
-                        <CRow className="mb-3">
-                            <CFormLabel htmlFor="userList" className="col-sm-2 col-form-label">User List</CFormLabel>
-                            <CCol md={6}>
-                                <CFormSelect id="userList" name="userList" multiple value={selectedGroup ? selectedGroup.user_set : []} onChange={handleUserSetChange}>
-                                    {userData.map(user => (
+                        <CFormLabel htmlFor="groupName">Group Name</CFormLabel>
+                        <CFormInput
+                            id="groupName"
+                            value={selectedGroup ? selectedGroup.name : ''}
+                            onChange={(e) => setSelectedGroup(prevGroup => ({ ...prevGroup, name: e.target.value }))}
+                        />
+                        <CFormLabel htmlFor="userSet" className="mt-3">Select Users</CFormLabel>
+                        <CFormSelect
+                            id="userSet"
+                            multiple
+                            value={selectedGroup ? selectedGroup.user_set : []}
+                            onChange={handleUserSetChange}
+                        >
+                            {userData.map(user => (
                                         <option key={user.usermod.id} value={user.usermod.id}>
                                             {user.usermod.username}
                                         </option>
                                     ))}
                                 </CFormSelect>
-                            </CCol>
-                        </CRow>
-                        <CRow className="justify-content-center">
-                            <CCol md="auto">
-                                <CButton color="primary" onClick={handleUpdateGroup}>Update</CButton>
-                            </CCol>
-                        </CRow>
                     </CForm>
                 </CModalBody>
                 <CModalFooter>
-                    <CButton color="secondary" onClick={() => setModalVisible(false)}>
-                        Close
+                    <CButton color="secondary" size='sm' onClick={() => setModalVisible(false)}>Close</CButton>
+                    <CButton color="primary" variant='outline' size='sm' onClick={handleUpdateGroup}>
+                        {selectedGroup && selectedGroup.id ? 'Update' : 'Create'}
                     </CButton>
                 </CModalFooter>
             </CModal>
-            <CModal size='lg' visible={newGroupModalVisible} onClose={() => setNewGroupModalVisible(false)} backdrop="static" keyboard={false}>
-                <CModalHeader onClose={() => setNewGroupModalVisible(false)}>
+            <CModal visible={newGroupModalVisible} onClose={() => setNewGroupModalVisible(false)}>
+                <CModalHeader>
                     <CModalTitle>Create New Group</CModalTitle>
                 </CModalHeader>
                 <CModalBody>
                     <CForm>
-                        <CRow className="mb-3">
-                            <CFormLabel htmlFor="newGroupName" className="col-sm-2 col-form-label">Group Name</CFormLabel>
-                            <CCol md={6}>
-                                <CFormInput type="text" id="newGroupName" name="newGroupName" value={newGroupName} onChange={(e) => setNewGroupName(e.target.value)} />
-                            </CCol>
-                        </CRow>
-                        <CRow className="mb-3">
-                            <CFormLabel htmlFor="newUserList" className="col-sm-2 col-form-label">User List</CFormLabel>
-                            <CCol md={6}>
-                                <CFormSelect id="newUserList" name="newUserList" multiple value={newGroupUsers} onChange={handleNewGroupUsersChange}>
-                                    {userData.map(user => (
+                        <CFormLabel htmlFor="newGroupName">Group Name</CFormLabel>
+                        <CFormInput
+                            id="newGroupName"
+                            value={newGroupName}
+                            onChange={(e) => setNewGroupName(e.target.value)}
+                        />
+                        <CFormLabel htmlFor="newGroupUsers" className="mt-3">Select Users</CFormLabel>
+                        <CFormSelect
+                            id="newGroupUsers"
+                            multiple
+                            value={newGroupUsers}
+                            onChange={handleNewGroupUsersChange}
+                        >
+                            {userData.map(user => (
                                         <option key={user.usermod.id} value={user.usermod.id}>
                                             {user.usermod.username}
                                         </option>
                                     ))}
                                 </CFormSelect>
-                            </CCol>
-                        </CRow>
-                        <CRow className="justify-content-center">
-                            <CCol md="auto">
-                                <CButton color="primary" onClick={handleCreateNewGroup}>Create</CButton>
-                            </CCol>
-                        </CRow>
                     </CForm>
                 </CModalBody>
                 <CModalFooter>
-                    <CButton color="secondary" onClick={() => setNewGroupModalVisible(false)}>
-                        Close
-                    </CButton>
+                    <CButton color="secondary" size='sm' onClick={() => setNewGroupModalVisible(false)}>Close</CButton>
+                    <CButton color="primary" variant='outline' size='sm' onClick={handleCreateNewGroup}>Create</CButton>
                 </CModalFooter>
             </CModal>
         </>
